Refactored common code of OpenAIChat and AzureOpenAIChat into class OpenAIBase
